Move (include|src)/runtime/InputCallback.* to (include|src)/input/InputCallback.*
Issue #14
